requirements.txt
fastapi>=0.109
uvicorn>=0.27
httpx>=0.27
pydantic>=2.5
google-cloud-aiplatform>=1.38
google-cloud-firestore>=2.14
google-cloud-storage>=2.14
google-cloud-pubsub>=2.18
google-cloud-dlp>=3.15
google-cloud-logging>=3.9
opentelemetry-api>=1.22
opentelemetry-sdk>=1.22
opentelemetry-exporter-gcp-trace>=1.6
pytest>=7.4
